Refactor project to align with Domain-Driven Design Rearranged the directory structure to separate concerns into Presentation, Application, Domain, and Infrastructure layers, following DDD best practices. Also added the Mapster library for efficient object-to-object mapping.
