+++ Common/Dtos/Common/OrderItemDto.cs
@@ -1,0 +1,12 @@
+﻿namespace OrderService.Application.Dtos.Request;
+
+public class OrderItemDto
+{
+    public int ProductId { get; set; } 
+    
+    public string ProductName { get; set; } 
+    
+    public int Quantity { get; set; } 
+    
+    public decimal UnitPrice { get; set; } 
+}
