--- Common/Dtos/OrderItemSnapshot.cs
@@ -1,10 +1,0 @@
-﻿namespace Common.Dtos;
-
-public class OrderItemSnapshot
-{
-    public int Productid { get; set; }
-    
-    public string ProductName {set;get;}
-    
-    public int Quantity {set;get;}
-}
