+++ OrderService/Application/Commands/CreateOrder/CreateOrderCommand.cs
@@ -1,0 +1,7 @@
+﻿using MediatR;
+using OrderService.Application.Dtos.Request;
+using OrderService.Domain.Agregates;
+
+namespace OrderService.Application.Commands.CreateOrder;
+
+public record CreateOrderCommand(string UserId, CreateOrderDto Order) : IRequest<Order>;
