+++ OrderService/Application/Commands/UpdateOrderStatus/UpdateOrderStatusCommand.cs
@@ -1,0 +1,6 @@
+﻿using MediatR;
+using OrderService.Domain.Agregates;
+
+namespace OrderService.Application.Commands.UpdateOrderStatus;
+
+public record UpdateOrderStatusCommand(int OrderId, string OrderStatus) : IRequest<Order>;
