+++ OrderService/Application/Dtos/Request/CreateOrderDto.cs
@@ -1,0 +1,8 @@
+﻿namespace OrderService.Application.Dtos.Request;
+
+public class CreateOrderDto
+{
+    public List<OrderItemDto> Items { get; set; } = [];
+    
+    public ShippingInfoDto Shipping { get; set; }
+}
