+++ OrderService/Application/Mapping/DependencyInjection.cs
@@ -1,0 +1,19 @@
+﻿using System.Reflection;
+using Mapster;
+using MapsterMapper;
+
+namespace OrderService.Application.Mapping;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddMapping(this IServiceCollection services)
+    {
+        var config = TypeAdapterConfig.GlobalSettings;
+        config.Scan(Assembly.GetExecutingAssembly());
+        
+        services.AddSingleton(config);
+        services.AddScoped<IMapper, ServiceMapper>();
+        
+        return services;
+    }
+}
