+++ OrderService/Application/Queries/GetOrderById/GetOrderByIdQuery.cs
@@ -1,0 +1,6 @@
+﻿using MediatR;
+using OrderService.Domain.Agregates;
+
+namespace OrderService.Application.Queries.GetOrderById;
+
+public record GetOrderByIdQuery(int OrderId) : IRequest<Order>;
