+++ OrderService/Application/Queries/GetOrders/GetOrdersQuery.cs
@@ -1,0 +1,6 @@
+﻿using MediatR;
+using OrderService.Domain.Agregates;
+
+namespace OrderService.Application.Queries.GetOrders;
+
+public record GetOrdersQuery() : IRequest<List<Order>>;
