+++ OrderService/Application/Queries/GetOrdersByUserId/GetOrdersByUserIdQuery.cs
@@ -1,0 +1,6 @@
+﻿using MediatR;
+using OrderService.Domain.Agregates;
+
+namespace OrderService.Application.Queries.GetOrdersByUserId;
+
+public record GetOrdersByUserIdQuery(string userId) : IRequest<List<Order>>;
