--- OrderService/Dtos/Request/CreateOrderDto.cs
@@ -1,10 +1,0 @@
-﻿using OrderService.Models;
-
-namespace OrderService.Dtos.Request;
-
-public class CreateOrderDto
-{
-    public List<CreateOrderItemDto> Items { get; set; } = [];
-    
-    public ShippingInfoDto Shipping { get; set; }
-}
