--- OrderService/Dtos/Request/CreateOrderItemDto.cs
@@ -1,12 +1,0 @@
-﻿namespace OrderService.Dtos.Request;
-
-public class CreateOrderItemDto
-{
-    public int ProductId { get; set; } 
-    
-    public string ProductName { get; set; } 
-    
-    public int Quantity { get; set; } 
-    
-    public decimal UnitPrice { get; set; } 
-}
