+++ OrderService/Infrastructure/Data/IOrderRepository.cs
@@ -1,0 +1,13 @@
+﻿using OrderService.Domain.Agregates;
+
+namespace OrderService.Infrastructure.Data;
+
+public interface IOrderRepository
+{
+    public Task<Order> GetOrderById(int orderId);
+    public Task<List<Order>> GetOrders();
+    public Task<List<Order>> GetOrdersByUserId(string userId);
+    public Task CreateOrder(Order order);
+    public Task UpdateOrder(Order order);
+
+}
