--- OrderService/OrderManagement/Command/UpdateOrderStatus/UpdateOrderStatusCommand.cs
@@ -1,7 +1,0 @@
-﻿using Common.Enums;
-using MediatR;
-using OrderService.Models;
-
-namespace OrderService.OrderManagement.Command.UpdateOrderStatus;
-
-public record UpdateOrderStatusCommand(int OrderId, string OrderStatus) : IRequest<Order>;
