--- OrderService/OrderManagement/Query/GetOrderById/GetOrderByIdQuery.cs
@@ -1,6 +1,0 @@
-﻿using MediatR;
-using OrderService.Models;
-
-namespace OrderService.OrderManagement.Query.GetOrderById;
-
-public record GetOrderByIdQuery(int OrderId) : IRequest<Order>;
