--- OrderService/OrderManagement/Query/GetOrders/GetOrdersQuery.cs
@@ -1,6 +1,0 @@
-﻿using MediatR;
-using OrderService.Models;
-
-namespace OrderService.OrderManagement.Query.GetOrders;
-
-public record GetOrdersQuery() : IRequest<List<Order>>;
