--- OrderService/OrderManagement/Query/GetOrdersByUserId/GetOrdersByUserIdQuery.cs
@@ -1,6 +1,0 @@
-﻿using MediatR;
-using OrderService.Models;
-
-namespace OrderService.OrderManagement.Query.GetOrdersByUserId;
-
-public record GetOrdersByUserIdQuery(string userId) : IRequest<List<Order>>;
